+++ CHANGELOG.md
@@ -14,7 +14,8 @@
 * Added a text instance of the carry weight range the character is in under the Carry Weight Bar.
 * Added a Recharge Shield button to replenish current Shield Integrity by the Recharge value, up to the Shield's maximum Integrity.
 * Fixed various spacing issues on all sheets.
-* Added the previously missing Flashbang, Smoke Grenade, and Tear Gas special rules to the Weapon Item sheet.
+* Added the Cryo, Flashbang, Smoke Grenade, and Tear Gas special rules to the Weapon Item sheet.
+* The Cryo, Flame, and Needle special rules now roll their damage in the attack chat message based on the number of hits the attacker landed.
 * Updated the Weapon Item sheet in preparation for a new data model that will be able to support Special Ammo down the line.
 
 ### Bug Fixes